[lk][trusty] Update structure member len form size_t to uint32_t
Update structure member len in struct iover_kern_t, iovec_user_t
and ipc_msg_info form size_t to uint32_t

+++ lib/trusty/include/lib/trusty/uio.h
@@ -29,12 +29,12 @@
 
 typedef struct iovec_kern {
 	void		*base;
-	size_t		len;
+	uint32_t		len;
 } iovec_kern_t;
 
 typedef struct iovec_user {
 	user_addr_t	base;
-	size_t	len;
+	uint32_t	len;
 } iovec_user_t;
